Add files via upload
Tower of Hanoi
This repository contains an implementation of the Tower of Hanoi puzzle. The Tower of Hanoi is a classic mathematical problem involving three pegs and a number of disks of different sizes. The objective is to move all the disks from the source peg to the destination peg using an auxiliary peg, adhering to the following rules:

Only one disk can be moved at a time.
A disk can only be placed on top of a larger disk or on an empty peg.
All disks start on the source peg in decreasing size, with the largest disk at the bottom.
Features
Recursive and iterative solutions
Visualization of the moves
Customizable number of disks
Step-by-step move instructions